`fix:` duas tabelas para usuário
