+++ database/factories/UserFactory.php
@@ -10,35 +10,41 @@ use Illuminate\Support\Str;
  */
 class UserFactory extends Factory
 {
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
-    public function definition()
-    {
-        return [
-            'name' => $this->faker->name(),
-            'email' => $this->faker->unique()->safeEmail(),
-            'email_verified_at' => now(),
-            'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
-            'remember_token' => Str::random(10),
-            /* put column user and admin with the expect value?*/
-            
-        ];
-    }
+	/**
+	 * Define the model's default state.
+	 *
+	 * @return array<string, mixed>
+	 */
+	public function definition()
+	{
+		return [
+			'first_name' => $this->faker->firstName(),
+			'last_name' => $this->faker->lastName(),
+			'birth_date' => $this->faker->date(),
+			'phone' => $this->faker->phoneNumber(),
+			'address' => $this->faker->address(),
 
-    /**
-     * Indicate that the model's email address should be unverified.
-     *
-     * @return static
-     */
-    public function unverified()
-    {
-        return $this->state(function (array $attributes) {
-            return [
-                'email_verified_at' => null,
-            ];
-        });
-    }
+			'cpf' => $this->faker->unique()->numberBetween(10000000000, 99999999999),
+			'email' => $this->faker->unique()->safeEmail(),
+
+			'email_verified_at' => now(),
+			'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
+			'remember_token' => Str::random(10),
+			/* put column user and admin with the expect value?*/
+		];
+	}
+
+	/**
+	 * Indicate that the model's email address should be unverified.
+	 *
+	 * @return static
+	 */
+	public function unverified()
+	{
+		return $this->state(function (array $attributes) {
+			return [
+				'email_verified_at' => null,
+			];
+		});
+	}
 }
