+++ database/seeders/DatabaseSeeder.php
@@ -2,29 +2,31 @@
 
 namespace Database\Seeders;
 
-use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
 {
-    /**
-     * Seed the application's database.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        \App\Models\User::factory(10)->create();
+	/**
+	 * Seed the application's database.
+	 *
+	 * @return void
+	 */
+	public function run()
+	{
+		\App\Models\User::factory(10)->create();
 
-        \App\Models\Product::factory(20)->create();
+		\App\Models\Product::factory(20)->create();
 
-        \App\Models\Client::factory(10)->create();
+		\App\Models\Order::factory(5)->create();
 
-        \App\Models\Order::factory(5)->create();
-
-    // \App\Models\User::factory()->create([
-    //     'name' => 'Test User',
-    //     'email' => 'test@example.com',
-    // ]);
-    }
+		\App\Models\User::factory()->create([
+			'first_name' => 'Admin',
+			'last_name' => 'Admin',
+			'phone' => '(00) 0 0000-0000',
+			'cpf' => '000.000.000-00',
+			'email' => 'admin@email.com',
+			'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
+			'admin' => true
+		]);
+	}
 }
